// src/app/shared/models/product.model.ts
export interface Product {
  id: number;
  name: string;
  sku: string;
  description: string;
  category: string;
  price: string;
  stock: boolean;
  currentStock: number;
  totalSales: number;
  totalRevenue: number;
  turnoverRate: string;
  revenue: number;
  growth: string;
  status: string;
  quickCode: string;
brandName: string;
quantityPack: string;
drugTypeName: string;
dosageFormName: string;
manufacturerName: string;
}